Updates modules to use ES6 module syntax.
