+++ lib/deferred.js
@@ -1,16 +1,9 @@
-(function (define) { 'use strict';
-  define(function () {
-
-    return function deferred() {
-      var d = {};
-      var promise = new Promise(function (resolve, reject) {
-        d.resolve = resolve;
-        d.reject = reject;
-      });
-      d.promise = promise;
-      return d;
-    };
-
+export default function deferred() {
+  var d = {};
+  var promise = new Promise(function (resolve, reject) {
+    d.resolve = resolve;
+    d.reject = reject;
   });
-})(typeof define === 'function' && define.amd ? define :
-/* istanbul ignore next */ function (factory) { module.exports = factory(require); });
+  d.promise = promise;
+  return d;
+}
