+++ lib/serialize_error.js
@@ -1,26 +1,19 @@
-(function (define) { 'use strict';
-  define(function () {
+export default function serializeError(error) {
+  var serializedError = {};
 
-    return function serializeError(error) {
-      var serializedError = {};
+  serializedError.name = error.name;
+  serializedError.message = error.message;
 
-      serializedError.name = error.name;
-      serializedError.message = error.message;
+  // IE
+  serializedError.number = error.number;
+  serializedError.description = error.description;
 
-      // IE
-      serializedError.number = error.number;
-      serializedError.description = error.description;
+  // Firefox
+  serializedError.fileName = error.fileName;
+  serializedError.lineNumber = error.lineNumber;
 
-      // Firefox
-      serializedError.fileName = error.fileName;
-      serializedError.lineNumber = error.lineNumber;
+  // Chrome / Firefox / latest IE
+  serializedError.stack = error.stack;
 
-      // Chrome / Firefox / latest IE
-      serializedError.stack = error.stack;
-
-      return serializedError;
-    };
-
-  });
-})(typeof define === 'function' && define.amd ? define :
-/* istanbul ignore next */ function (factory) { module.exports = factory(require); });
+  return serializedError;
+}
